+++ CHANGELOG.md
@@ -7,3 +7,14 @@ Get Mobile Data Analysis
 ## 0.0.2
 
 Updated Affordabilily Analysis to include optional averageMonthlyTotalExpenses and averageMonthlyLoanRepaymentAmount in its payload
+
+## 0.0.3 - 2022-05-14
+
+Update the package with new features:
+* Generate Credit Score 
+* Get Existing Credit Score
+* Get Statement Transaction 
+* Get Statement Analysis
+* Get Existing Affordabilit Statement
+* Add Clients Identification to Statement
+